# openapi/user-service.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8083/api
  description: Generated server url
...
paths:
...
  /v1/budget-package-offer/deactivate/{id}:
    put:
      tags:
      - Budget package offer API
      description: Returns deactivated package by it's id.
      operationId: deactivateBudgetOfferPackage
      parameters:
      - name: id
        in: path
        description: UUID of desired package
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "400":
          description: Path parameter does not have a form of UUID
          content:
            application/json:
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "404":
          description: Desired package does not exist
          content:
            application/json:
              schema:
                type: string
        "200":
          description: Budget package was deactivated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetOfferPackageDto"
  /v1/budget-package-offer/activate/{id}:
    put:
      tags:
      - Budget package offer API
      description: Returns activated package by it's id.
      operationId: activateBudgetOfferPackage
      parameters:
      - name: id
        in: path
        description: UUID of desired package
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "400":
          description: Path parameter does not have a form of UUID
          content:
            application/json:
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "404":
          description: Desired package does not exist
          content:
            application/json:
              schema:
                type: string
        "200":
          description: Budget package was activated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetOfferPackageDto"
  /v1/budget-package-offer/:
    post:
      tags:
      - Budget package offer API
      description: Creates a package with given properties
      operationId: createBudgetOfferPackage
      requestBody:
        description: Data for pcakage creation
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetOfferPackageDto"
        required: true
      responses:
        "400":
          description: Validation of input data failed
          content:
            application/json:
              schema:
                type: string
        "409":
          description: Entity already exists
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "201":
          description: Package with newly created package data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetOfferPackageDto"
...
  /v1/budget-package-offer/{id}:
    get:
      tags:
      - Budget package offer API
      description: Returns desired budget package by its id.
      operationId: getBudgetOfferPackageById
      parameters:
      - name: id
        in: path
        description: UUID of desired package
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "400":
          description: Path parameter does not have a form of UUID
          content:
            application/json:
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "404":
          description: Desired package does not exist
          content:
            application/json:
              schema:
                type: string
        "200":
          description: Budget package with desired details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetOfferPackageDto"
  /v1/budget-package-offer/available-packages:
    get:
      tags:
      - Budget package offer API
      description: Returns all available budget packages.
      operationId: getAllAvailableBudgetOfferPackages
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: List of all available packages.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetOfferPackageDto"
  /v1/budget-package-offer/all-packages:
    get:
      tags:
      - Budget package offer API
      description: Returns all budget packages.
      operationId: getAllBudgetOfferPackages
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: List of all packages.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetOfferPackageDto"
components:
  schemas:
...
    BudgetOfferPackageDto:
      type: object
      properties:
        guid:
          type: string
          format: uuid
        priceDollars:
          type: integer
          format: int32
          minimum: 0
        budgetIncrease:
          type: integer
          format: int32
          minimum: 1
        description:
          type: string
        isAvailable:
          type: boolean
      required:
      - isAvailable
...